Бабушкина У.М. 25-ИСТ-2

+++ src/alg.cpp
@@ -1,4 +1,7 @@
 // Copyright 2022 NNTU-CS
+#include <cstdint>
+#include "alg.h"
+
 double pow(double value, uint16_t n) {
     double result = 1.0;
     for (uint16_t i = 0; i < n; ++i) {
@@ -17,7 +20,7 @@ uint64_t fact(uint16_t n) {
 
 double calcItem(double x, uint16_t n) {
     if (n == 0) return 1.0;
-    return pow(x, n) / fact(n);
+    return pow(x, n) / static_cast<double>(fact(n));
 }
 
 double expn(double x, uint16_t count) {
@@ -32,7 +35,7 @@ double sinn(double x, uint16_t count) {
     double result = 0.0;
     for (uint16_t n = 1; n <= count; ++n) {
         uint16_t power = 2 * n - 1;
-        double term = pow(x, power) / fact(power);
+        double term = pow(x, power) / static_cast<double>(fact(power));
         if (n % 2 == 0) {
             result -= term;
         } else {
@@ -46,7 +49,7 @@ double cosn(double x, uint16_t count) {
     double result = 0.0;
     for (uint16_t n = 1; n <= count; ++n) {
         uint16_t power = 2 * (n - 1);
-        double term = pow(x, power) / fact(power);
+        double term = pow(x, power) / static_cast<double>(fact(power));
         if (n % 2 == 0) {
             result -= term;
         } else {
